Javaify class names
