+++ tapglue-android-sdk/src/main/java/com/tapglue/managers/TGConnectionManager.java
@@ -19,193 +19,24 @@ package com.tapglue.managers;
 
 import android.support.annotation.NonNull;
 
-import com.tapglue.Tapglue;
-import com.tapglue.model.TGConnection;
-import com.tapglue.model.TGConnection.TGConnectionState;
 import com.tapglue.model.TGConnection.TGConnectionType;
 import com.tapglue.model.TGPendingConnections;
 import com.tapglue.networking.requests.TGRequestCallback;
-import com.tapglue.networking.requests.TGRequestErrorType;
-import com.tapglue.networking.requests.TGRequestErrorType.ErrorType;
 
+public interface TGConnectionManager {
+    void confirmConnection(@NonNull Long userId, @NonNull TGConnectionType TGConnectionType, @NonNull final TGRequestCallback<Boolean> callback);
 
-public class TGConnectionManager extends AbstractTGManager implements TGConnectionManagerInterface {
+    void followUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback);
 
-    public TGConnectionManager(Tapglue tgInstance) {
-        super(tgInstance);
-    }
+    void friendUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback);
 
-    @Override
-    public void confirmConnection(@NonNull Long userId, @NonNull TGConnectionType connectionType, @NonNull final TGRequestCallback<Boolean> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().confirmConnection(userId, connectionType, new TGRequestCallback<TGConnection>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
+    void getPendingConnections(@NonNull final TGRequestCallback<TGPendingConnections> callback);
 
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
+    void retrieveConfirmedConnectionsForCurrentUser(@NonNull final TGRequestCallback<TGPendingConnections> callback);
 
-            @Override
-            public void onRequestFinished(TGConnection output, boolean changeDoneOnline) {
-                callback.onRequestFinished(true, true);
-            }
-        });
-    }
+    void retrieveRejectedConnectionsForCurrentUser(@NonNull final TGRequestCallback<TGPendingConnections> callback);
 
-    /**
-     * Follow selected user
-     *
-     * @param userId
-     * @param callback
-     */
-    @Override
-    public void followUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().createConnection(userId, TGConnectionType.FOLLOW, TGConnectionState.CONFIRMED, new TGRequestCallback<TGConnection>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
+    void unfollowUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback);
 
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
-
-            @Override
-            public void onRequestFinished(TGConnection output, boolean changeDoneOnline) {
-                callback.onRequestFinished(true, true);
-            }
-        });
-    }
-
-    /**
-     * Add selected user to friends
-     *
-     * @param userId   user ID
-     * @param callback return method
-     */
-    @Override
-    public void friendUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().createConnection(userId, TGConnectionType.FRIEND, TGConnectionState.PENDING, new TGRequestCallback<TGConnection>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
-
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
-
-            @Override
-            public void onRequestFinished(TGConnection output, boolean changeDoneOnline) {
-                callback.onRequestFinished(true, true);
-            }
-        });
-    }
-
-    /**
-     * Get pending Connections
-     */
-    @Override
-    public void getPendingConnections(@NonNull final TGRequestCallback<TGPendingConnections> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().createPendingConnectionsRequest(callback);
-    }
-
-    @Override
-    public void retrieveConfirmedConnectionsForCurrentUser(@NonNull final TGRequestCallback<TGPendingConnections> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().createConfirmedConnectionsRequest(callback);
-    }
-
-    @Override
-    public void retrieveRejectedConnectionsForCurrentUser(@NonNull final TGRequestCallback<TGPendingConnections> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().createRejectedConnectionsRequest(callback);
-    }
-
-    /**
-     * Stop following selected user
-     *
-     * @param userId   user ID
-     * @param callback return method
-     */
-    @Override
-    public void unfollowUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().removeConnection(userId, TGConnectionType.FOLLOW, new TGRequestCallback<Object>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
-
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
-
-            @Override
-            public void onRequestFinished(Object output, boolean changeDoneOnline) {
-                callback.onRequestFinished(true, true);
-            }
-        });
-    }
-
-    /**
-     * Remove user from friends
-     *
-     * @param callback return method
-     * @param userId   ID of user
-     */
-    @Override
-    public void unfriendUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().removeConnection(userId, TGConnectionType.FRIEND, new TGRequestCallback<Object>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
-
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
-
-            @Override
-            public void onRequestFinished(Object output, boolean changeDoneOnline) {
-                callback.onRequestFinished(true, true);
-            }
-        });
-    }
+    void unfriendUser(@NonNull Long userId, @NonNull final TGRequestCallback<Boolean> callback);
 }
