+++ tapglue-android-sdk/src/main/java/com/tapglue/managers/TGFeedManager.java
@@ -17,247 +17,44 @@
 
 package com.tapglue.managers;
 
-import android.content.Context;
-import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import com.tapglue.Tapglue;
 import com.tapglue.model.TGEventsList;
 import com.tapglue.model.TGFeed;
 import com.tapglue.model.TGFeedCount;
 import com.tapglue.model.TGPostsList;
 import com.tapglue.model.queries.TGQuery;
 import com.tapglue.networking.requests.TGRequestCallback;
-import com.tapglue.networking.requests.TGRequestErrorType;
 
-public class TGFeedManager extends AbstractTGManager implements TGFeedManagerInterface {
+public interface TGFeedManager {
+    void cachedFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback);
 
-    private static final String CACHE_KEY = "FEED_CACHE";
+    void getCachedFeedIfAvailable(@NonNull final TGRequestCallback<TGFeed> callback);
 
-    public TGFeedManager(Tapglue tgInstance) {
-        super(tgInstance);
-    }
+    void retrieveEventsFeedForCurrentUser(@NonNull final TGRequestCallback<TGEventsList> callback);
 
-    /**
-     * Get feed from cache
-     *
-     * @param callback
-     */
-    @Override
-    public void cachedFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        getCachedFeedIfAvailable(callback);
-    }
+    void retrieveEventsFeedForCurrentUser(@Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback);
 
-    /**
-     * Return cached feed
-     *
-     * @param callback
-     */
-    @Override
-    public void getCachedFeedIfAvailable(@NonNull final TGRequestCallback<TGFeed> callback) {
-        SharedPreferences cache = tapglue.getContext().getSharedPreferences(TGFeedManager.class.toString(), Context.MODE_PRIVATE);
-        if (!cache.contains(CACHE_KEY)) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.NO_CACHE_OBJECT));
-            return;
-        }
+    void retrieveEventsForCurrentUser(@NonNull final TGRequestCallback<TGEventsList> callback);
 
-        TGFeed feed = new Gson().fromJson(cache.getString(CACHE_KEY, null), new TypeToken<TGFeed>() {}.getType());
-        callback.onRequestFinished(feed, false);
-    }
+    void retrieveEventsForCurrentUser(@Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback);
 
-    /**
-     * Get all events and posts associated with current user
-     *
-     * @param callback
-     */
-    @Override
-    public void retrieveEventsFeedForCurrentUser(@NonNull final TGRequestCallback<TGEventsList> callback) {
-        retrieveEventsFeedForCurrentUser(null, callback);
-    }
+    void retrieveEventsForUser(@NonNull Long userId, @NonNull final TGRequestCallback<TGEventsList> callback);
 
-    @Override
-    public void retrieveEventsFeedForCurrentUser(@Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getEvents(whereParameters, callback);
-    }
+    void retrieveEventsForUser(@NonNull Long userId, @Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback);
 
-    /**
-     * Get all events associated with current user Warning, Posts in TGEventsList output object will
-     * be empty as this method returns only events!
-     *
-     * @param callback
-     */
-    @Override
-    public void retrieveEventsForCurrentUser(@NonNull final TGRequestCallback<TGEventsList> callback) {
-        retrieveEventsForCurrentUser(null, callback);
-    }
+    void retrieveNewsFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback);
 
-    @Override
-    public void retrieveEventsForCurrentUser(@Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback) {
+    void retrieveNewsFeedForCurrentUser(@NonNull TGQuery whereParameters, @NonNull final TGRequestCallback<TGFeed> callback);
 
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getEvents(whereParameters, callback);
-    }
+    void retrievePostsFeedForCurrentUser(@NonNull final TGRequestCallback<TGPostsList> callback);
 
-    /**
-     * Get all events associated with user Warning, Posts in TGEventsList output object will be
-     * empty as this method returns only events!
-     *
-     * @param userId
-     * @param callback
-     */
-    @Override
-    public void retrieveEventsForUser(@NonNull Long userId, @NonNull final TGRequestCallback<TGEventsList> callback) {
-        retrieveEventsForUser(userId, null, callback);
-    }
+    void retrievePostsForCurrentUser(@NonNull final TGRequestCallback<TGPostsList> callback);
 
-    @Override
-    public void retrieveEventsForUser(@NonNull Long userId, TGQuery whereParameters, @NonNull final TGRequestCallback<TGEventsList> callback) {
-        if (userId == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.NULL_INPUT));
-            return;
-        }
-        else if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getEvents(userId, whereParameters, callback);
-    }
+    void retrievePostsForUser(@NonNull Long userId, @NonNull final TGRequestCallback<TGPostsList> callback);
 
-    /**
-     * Get live feed for current user
-     *
-     * @param callback
-     */
-    @Override
-    public void retrieveNewsFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback) {
-        retrieveNewsFeedForCurrentUser(null, callback);
-    }
+    void retrieveUnreadCountForCurrentUser(@NonNull final TGRequestCallback<TGFeedCount> callback);
 
-    @Override
-    public void retrieveNewsFeedForCurrentUser(@Nullable TGQuery whereParameters, @NonNull final TGRequestCallback<TGFeed> callback) {
-        tapglue.createRequest().getFeed(whereParameters, new TGRequestCallback<TGFeed>() {
-            @Override
-            public boolean callbackIsEnabled() {
-                return callback.callbackIsEnabled();
-            }
-
-            @Override
-            public void onRequestError(TGRequestErrorType cause) {
-                callback.onRequestError(cause);
-            }
-
-            @Override
-            public void onRequestFinished(TGFeed output, boolean changeDoneOnline) {
-                saveFeedToCache(output);
-                callback.onRequestFinished(output, changeDoneOnline);
-            }
-        });
-    }
-
-    /**
-     * Get all events and posts associated with current user
-     *
-     * @param callback
-     */
-    @Override
-    public void retrievePostsFeedForCurrentUser(@NonNull final TGRequestCallback<TGPostsList> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getMyPosts(callback);
-    }
-
-    /**
-     * Get all posts associated with current user
-     *
-     * @param callback
-     */
-    @Override
-    public void retrievePostsForCurrentUser(@NonNull final TGRequestCallback<TGPostsList> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getMyPosts(callback);
-    }
-
-    /**
-     * Get all posts associated with user
-     *
-     * @param userId
-     * @param callback
-     */
-    @Override
-    public void retrievePostsForUser(@NonNull Long userId, @NonNull final TGRequestCallback<TGPostsList> callback) {
-        if (userId == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.NULL_INPUT));
-            return;
-        }
-        else if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getUserPosts(userId, callback);
-    }
-
-    /**
-     * Get count of unread feed elements
-     *
-     * @param callback
-     */
-    @Override
-    public void retrieveUnreadCountForCurrentUser(@NonNull final TGRequestCallback<TGFeedCount> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getFeedCount(callback);
-    }
-
-    /**
-     * Get all unread feed
-     *
-     * @param callback
-     */
-    @Override
-    public void retrieveUnreadFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback) {
-        if (tapglue.getUserManager().getCurrentUser() == null) {
-            callback.onRequestError(new TGRequestErrorType(TGRequestErrorType.ErrorType.USER_NOT_LOGGED_IN));
-            return;
-        }
-        tapglue.createRequest().getUnreadFeed(callback);
-    }
-
-    /**
-     * Save feed to cache
-     *
-     * @param output
-     */
-    private void saveFeedToCache(@Nullable TGFeed output) {
-        SharedPreferences cache = tapglue.getContext().getSharedPreferences(TGFeedManager.class.toString(), Context.MODE_PRIVATE);
-        if (output == null) {
-            if (cache.contains(CACHE_KEY)) {
-                cache.edit().remove(CACHE_KEY).apply();
-            }
-        }
-        else {
-            cache.edit().putString(CACHE_KEY, new Gson().toJson(output, new TypeToken<TGFeed>() {
-            }.getType())).apply();
-        }
-    }
+    void retrieveUnreadFeedForCurrentUser(@NonNull final TGRequestCallback<TGFeed> callback);
 }
